agregado boton de confirmar

+++ public/js/education.js
@@ -77,6 +77,10 @@ function formatDate(dateObj) {
 }
 
 function deleteEducation(id, button) {
+    if (!confirm("¿Estás seguro de que deseas eliminar esta educación?")) {
+        return; // Detiene la ejecución si el usuario cancela
+    }
+
     const deleteUrl = button.getAttribute("data-url");
     fetch(deleteUrl, {
         method: 'DELETE', // Asegura que el backend acepte DELETE, si no usa 'POST'
